derniere maj

+++ app/Http/Controllers/PostController.php
@@ -4,20 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
+use Illuminate\View\View; 
+use Illuminate\Http\RedirectResponse;
 
 class PostController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(): View
     {
-        //On récupère tous les Post
-        $posts = Post::latest()->get();
-
-        // On transmet les Post à la vue
-        return view("posts.index", compact("posts"));
+        return view('posts.index', [
+            'posts' => Post::with('user')->latest()->get(),
+        ]);
     }
 
     /**
@@ -25,37 +24,29 @@ class PostController extends Controller
      */
     public function create()
     {
-        // On retourne la vue "/resources/views/posts/edit.blade.php"
-        return view("posts.edit");
+        return view('posts.create');
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(Request $request): RedirectResponse
     {
-        // 1. La validation
-        $this->validate($request, [
-            'title' => 'bail|required|string|max:255',
-            "picture" => 'bail|required|image|max:1024',
-            "content" => 'bail|required',
+        $validated = $request->validate([
+            'content' => 'required|string|max:255',
+            'title' => 'required|string|max:255',
+            'image' => 'required|image'
         ]);
         
-        
-        // 2. On upload l'image dans "/storage/app/public/posts"
-        $chemin_image = $request->picture->store("posts"); 
+        $validated = [
+            'title' => $request->title,
+            'content' => $request->content,
+            'image' => $request->image->store('posts')
+        ];
+
+        $request->user()->posts()->create($validated);
  
-
-        // 3. On enregistre les informations du Post
-        Post::create([
-            "title" => $request->title,
-            "picture" => $chemin_image,
-            "content" => $request->content,
-        ]);
-
-
-        // 4. On retourne vers tous les posts : route("posts.index")
-        return redirect(route("posts.index"));
+        return redirect(route('posts.index'));
     }
 
     /**
@@ -63,70 +54,47 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return view("posts.show", compact("post"));
+        return view('posts.show', compact('post'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Post $post)
+    public function edit(Post $post): View
     {
-        return view("posts.edit", compact("post"));
+
+        $this->authorize('update', $post);
+ 
+        return view('posts.edit', [
+            'posts' => $post,
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Post $post)
+    public function update(Request $request, Post $post): RedirectResponse
     {
-        // 1. La validation
-
-        // Les règles de validation pour "title" et "content"
-        $rules = [
-            'title' => 'bail|required|string|max:255',
-            "content" => 'bail|required',
-        ];
-
-        // Si une nouvelle image est envoyée
-        if ($request->has("picture")) {
-            // On ajoute la règle de validation pour "picture"
-            $rules["picture"] = 'bail|required|image|max:1024';
-        }
-
-        $this->validate($request, $rules);
-
-        // 2. On upload l'image dans "/storage/app/public/posts"
-        if ($request->has("picture")) {
-
-            //On supprime l'ancienne image
-            Storage::delete($post->picture);
-
-            $chemin_image = $request->picture->store("posts");
-        }
-
-        // 3. On met à jour les informations du Post
-        $post->update([
-            "title" => $request->title,
-            "picture" => isset($chemin_image) ? $chemin_image : $post->picture,
-            "content" => $request->content
+        $this->authorize('update', $post);
+        $validated = $request->validate([
+            'content' => 'required|string|max:255',
+            'title' => 'required|string|max:255',
+            'image' => 'required|image'
         ]);
-
-        // 4. On affiche le Post modifié : route("posts.show")
-        return redirect(route("posts.show", $post));
+        
+        $post->update($validated);
+        return redirect(route('posts.index'));
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Post $post)
+    public function destroy(Post $post):RedirectResponse
     {
-        // On supprime l'image existant
-        Storage::delete($post->picture);
-
-        // On les informations du $post de la table "posts"
+        $this->authorize('delete', $post);
+ 
         $post->delete();
-
-        // Redirection route "posts.index"
-        return redirect(route('posts.index'));
+ 
+        return redirect(route('posts.index')); 
     }
 }
